requirements.txt
librosa==0.10.2.post1
openai-whisper==20240313
nltk==3.9.1
scikit-learn==1.5.1
matplotlib==3.9.2
seaborn==0.13.2
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
gunicorn==22.0.0
Flask==3.0.3